Feat : Formulaire Repondant & Entreprise DONE

+++ views/employer/list.php
@@ -4,7 +4,7 @@
 
     <title>Liste des Administrateurs</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-<!--    <link rel="stylesheet" href="https://cdn.datatables.net/1.11.5/css/jquery.dataTables.min.css">-->
+
 </head>
 
 <body>
@@ -20,7 +20,7 @@ include dirname(__DIR__).'/../navbarEmployer.php';
         </div>
 <!--        <button type="button" class="btn btn-outline-secondary">Secondary</button>-->
         <div class="card-body">
-            <a href="http://localhost/PHP/NousLesFemmes/employer/create" class="btn btn-outline-secondary">Ajouter un employer</a><br><br>
+            <a href="employer/create" class="btn btn-outline-secondary">Ajouter un employer</a><br><br>
             <table id="myTable" class="table table-bordered">
                 <tr>
                     <th>Id</th>
@@ -41,7 +41,7 @@ include dirname(__DIR__).'/../navbarEmployer.php';
 
                         <td>
                             <a href="edit.php?id=<?= $employer->id?>" class="btn btn-outline-warning">Modifier</a>
-                            <a href="./employer/delete/<?= $employer->id?>" class="btn btn-outline-danger">Supprimer</a>
+                            <a onclick="return confirm('Voulez-vous supprimer cet employer')" href="employer/delete/<?= $employer->id?>" class="btn btn-outline-danger">Supprimer</a>
                         </td>
                     </tr>
                 <?php endforeach?>
